+++ cluster/webui.lua
@@ -297,13 +297,40 @@ local gql_type_replicaset = gql_types.object {
     name = 'Replicaset',
     description = 'Group of servers replicating the same data',
     fields = {
-        uuid = gql_types.string.nonNull,
-        roles = gql_types.list(gql_types.string.nonNull),
-        status = gql_types.string.nonNull,
-        weight = gql_types.float,
-        master = gql_types.nonNull('Server'),
-        active_master = gql_types.nonNull('Server'),
-        servers = gql_types.list('Server'),
+        uuid = {
+            kind = gql_types.string.nonNull,
+            description = 'The replica set uuid',
+        },
+        roles = {
+            kind = gql_types.list(gql_types.string.nonNull),
+            description = 'The role set enabled' ..
+                ' on every instance in the replica set',
+        },
+        status = {
+            kind = gql_types.string.nonNull,
+            description = 'The replica set health.' ..
+                ' It is "healthy" if all instances have status "healthy".' ..
+                ' Otherwise "unhealthy".',
+        },
+        weight = {
+            kind = gql_types.float,
+            description = 'Vshard replica set weight.' ..
+                ' Null for replica sets with vshard-storage role disabled.'
+        },
+        master = {
+            kind = gql_types.nonNull('Server'),
+            description = 'The leader according to the configuration.',
+        },
+        active_master = {
+            kind = gql_types.nonNull('Server'),
+            description = 'The active leader. It may differ from' ..
+                ' "master" if failover is enabled and configured leader' ..
+                ' isn\'t healthy.'
+        },
+        servers = {
+            kind = gql_types.list(gql_types.nonNull('Server')).nonNull,
+            description = 'Servers in the replica set.'
+        },
     }
 }
 
@@ -317,6 +344,10 @@ local gql_type_server = gql_types.object {
         status = gql_types.string.nonNull,
         message = gql_types.string.nonNull,
         disabled = gql_types.boolean.nonNull,
+        priority = {
+            kind = gql_types.int.nonNull,
+            description = 'Failover priority within the replica set',
+        },
         replicaset = gql_type_replicaset,
         boxinfo = boxinfo_schema,
         statistics = statistics_schema,
@@ -543,7 +574,7 @@ local function init(httpd)
         args = {
             uuid = gql_types.string.nonNull,
             roles = gql_types.list(gql_types.string.nonNull),
-            master = gql_types.string,
+            master = gql_types.list(gql_types.string.nonNull),
             weight = gql_types.float,
         },
         kind = gql_types.boolean,