+++ webui/src/components/ReplicasetEditModal/ReplicasetEditModal.js
@@ -1,3 +1,4 @@
+import * as R from 'ramda';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { defaultMemoize } from 'reselect';
@@ -15,8 +16,21 @@ const styles = {
     font-size: 14px;
     color: #343434;
     font-family: Roboto;
+  `,
+  dragIcon: css`
+    width: 100%;
+    cursor: move!important;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: lightgrey;
+    font-size: 20px;
+  `,
+  tableStyles: css`
+    margin-left: 0px
   `
-}
+};
 
 
 const isStorageWeightInputDisabled = formData => ! formData.roles.includes('vshard-storage');
@@ -26,8 +40,63 @@ const isStorageWeightInputValueValid = formData => {
   return number >= 0 && number < Infinity;
 };
 
-const prepareFields = roles => {
+const renderSelectOptions = record => record.servers.map(server => ({
+  key: server.uuid,
+  label: <span className={styles.serverLabel}>
+            {server.alias || 'No alias'}{' '}
+    <span className={styles.uriLabel}>{server.uri}</span>
+          </span>,
+}));
+
+const renderDraggableListOptions = record => record.servers.map(server => ({
+  key: server.uuid,
+  label: <span className={styles.serverLabel}>
+            {server.alias || 'No alias'}{' '}
+    <span className={styles.uriLabel}>{server.uri}</span>
+          </span>,
+}));
+
+const prepareFields = (roles, replicaset) => {
   const rolesOptions = roles.map(role => ({ key: role, label: role }));
+
+  const shallRenderDraggableList = replicaset && replicaset.servers.length > 2;
+
+  const draggableListCustomProps = {};
+
+  if (!shallRenderDraggableList) {
+    draggableListCustomProps.create = {
+      hidden: true,
+    };
+  } else {
+    draggableListCustomProps.create = {
+      tableProps: {
+        showHeader: false,
+        className: styles.tableStyles,
+        rowKey: 'uuid',
+      },
+      tableColumns: [
+        {
+          title: 'Operates',
+          key: 'operate',
+          render: () => <a className={styles.dragIcon}>☰</a>,
+        },
+
+        {
+          title: 'Альяс',
+          dataIndex: 'alias',
+        },
+        {
+          title: 'Адрес',
+          dataIndex: 'uri',
+        },
+      ],
+      tableData: R.pipe(
+          R.map(R.pick(['alias', 'uri', 'uuid'])),
+          R.map((data) => ({ ...data, key: data.uuid })),
+      )(replicaset.servers)
+    };
+    draggableListCustomProps.edit = draggableListCustomProps.create;
+  }
 
   return [
     {
@@ -39,25 +108,6 @@ const prepareFields = roles => {
       title: 'Roles',
       type: 'checkboxGroup',
       options: rolesOptions,
-    },
-    {
-      key: 'master',
-      title: 'Master',
-      type: 'optionGroup',
-      options: record => {
-        return record.servers.map(server => ({
-          key: server.uuid,
-          label: <span className={styles.serverLabel}>
-            {server.alias || 'No alias'}{' '}
-            <span className={styles.uriLabel}>{server.uri}</span>
-          </span>,
-        }));
-      },
-      customProps: {
-        create: {
-          hidden: true,
-        },
-      },
     },
     {
       key: 'weight',
@@ -77,6 +127,14 @@ const prepareFields = roles => {
       },
       invalidFeedback: 'Field accepts number, ex: 1.2',
     },
+    {
+      key: 'master',
+      title: shallRenderDraggableList ? 'Priority' : 'Master',
+      type: shallRenderDraggableList ? 'draggableList' : 'optionGroup',
+      options: shallRenderDraggableList ? renderDraggableListOptions : renderSelectOptions,
+      customProps: draggableListCustomProps,
+    },
+
   ];
 };
 
@@ -130,8 +188,8 @@ class ReplicasetEditModal extends React.PureComponent {
   };
 
   getFields = () => {
-    const { knownRoles } = this.props;
-    return this.prepareFields(knownRoles);
+    const { knownRoles, replicaset } = this.props;
+    return this.prepareFields(knownRoles, replicaset);
   };
 
   getDataSource = () => {