+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Failover priority configuration using WebUI
+
 ### Changed
 
 - Instance restart now triggers config validation before roles initialization